(MODULES-1947) Improve support for MongoDB authentication and replicaset management. Adds the ability to create an 'administration' MongoDB user account, which then gets stored in a mongorc.js file which enables Puppet to connect to MongoDB without credentials. Admin username and password can be over-ridden via 'admin_username' and 'admin_password' parameters. Replica set configuration can be completed as part of mongodb::server class by either providing a list of members using 'replset_members', or a full replica set config hash using 'replset_config'. Alternatively, mongodb::replset can be used to configure replicaset seperately. Any attempt to manage mongodb_db or mongodb_user resources on non-master replicaset members will generate a warning instead of failing.

+++ lib/puppet/provider/mongodb_database/mongodb.rb
@@ -26,11 +26,19 @@ Puppet::Type.type(:mongodb_database).provide(:mongodb, :parent => Puppet::Provid
   end
 
   def create
-    mongo_eval('db.dummyData.insert({"created_by_puppet": 1})', @resource[:name])
+    if db_ismaster
+      mongo_eval('db.dummyData.insert({"created_by_puppet": 1})', @resource[:name])
+    else
+      Puppet.warning 'Database creation is available only from master host'
+    end
   end
 
   def destroy
-    mongo_eval('db.dropDatabase()', @resource[:name])
+    if db_ismaster
+      mongo_eval('db.dropDatabase()', @resource[:name])
+    else
+      Puppet.warning 'Database removal is available only from master host'
+    end
   end
 
   def exists?
